Delete unused images and update code

+++ app/Http/Controllers/PostsController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use App\Http\Requests\PostStoreRequest;
+use App\Http\Requests\PostUpdateRequest;
 use App\Models\Post;
 
 class PostsController extends Controller
@@ -18,31 +20,25 @@ class PostsController extends Controller
    */
 
   // Método para mostrar una lista de posts
-  public function index(){
+  public function index()
+  {
     $posts = Post::all();
     return view('posts.index', compact('posts'));
   }
 
   // Método para guardar un nuevo post en la base de datos
-  public function store(Request $req){
-
+  public function store(PostStoreRequest $req)
+  {
     $post = new Post;
 
-    // Validamos los datos del request
-    $req->validate([
-      'title' => 'required|min:3|max:255',
-      'body' => 'required|min:3',
-    ]);
-    
-    if($req->hasFile('image')) {
-      // Obtener el archivo
+    if ($req->hasFile('image')) 
+    {
       $file = $req->file('image');
       $path = 'img/';
-      $fileName = time().'-'.$file->getClientOriginalName();
+      $fileName = time() . '-' . $file->getClientOriginalName();
       $uploadSuccess = $file->move($path, $fileName);
-      $post->image = $path.$fileName;
+      $post->image = $path . $fileName;
     }
-    
 
     $post->title = $req->title;
     $post->body = $req->body;
@@ -52,28 +48,21 @@ class PostsController extends Controller
   }
 
   // Método para mostrar el formulario de edición con los datos del post
-  public function edit($id){
+  public function edit($id)
+  {
     $post = Post::find($id);
     return view('posts.edit', compact('post'));
   }
 
   // Método para actualizar un post existente en la base de datos
-  public function update(Request $req, $id){
+  public function update(PostUpdateRequest $req, $id)
+  {
 
     $post = Post::find($id);
-    
-    // Validamos los datos del request
-    $req->validate([
-      'title' => 'required|min:3|max:255',
-      'body' => 'required|min:3',
-    ]);
 
     // Eliminamos la imagen anterior solo si se proporciona una nueva imagen
     if ($req->hasFile('image')) {
-      
-      if (file_exists($post->image)) {
-        unlink($post->image);
-      }
+      if (file_exists($post->image)) unlink($post->image);
 
       // Obtener el archivo
       $file = $req->file('image');
@@ -92,15 +81,30 @@ class PostsController extends Controller
   }
 
   // Método para eliminar un post existente de la base de datos
-  public function destroy($id){
+  public function destroy($id)
+  {
     $post = Post::find($id);
+
+    // Eliminamos la imagen del post
+    if (file_exists($post->image)) {
+      unlink($post->image);
+    }
     $post->delete();
-    return redirect()->route('home')->with('success', 'Post eliminado correctamente');
+
+    return redirect()->back()->with('success', 'Post eliminado correctamente');
   }
 
   // Método para mostrar un post específico
-  public function show($id){
+  public function show($id)
+  {
     $post = Post::find($id);
     return view('posts.show', compact('post'));
+  }
+
+  // Método para mostrar una tabla con los posts
+  public function table()
+  {
+    $posts = Post::all();
+    return view('posts.table', compact('posts'));
   }
 }
